refactor: chatwoot feature folder
