+++ clients/web/src/app.d.ts
@@ -5,9 +5,8 @@
 // for information about these interfaces
 
 import type {SSRData} from "@urql/core/dist/types/exchanges/ssr";
-import type {
-    Chatwoot, ChatwootSDK, Config, SessionUser,
-} from "./lib/utils";
+import type {Config, SessionUser} from "./lib/utils";
+import type {Chatwoot, ChatwootSDK} from "./lib/components/abstract/Chatwoot";
 
 declare global {
     declare namespace App {
