+++ clients/web/src/lib/components/abstract/Chatwoot/index.ts
@@ -1,0 +1,1 @@
+export {default as Chatwoot} from "./Chatwoot.svelte";
