+++ clients/web/src/lib/components/abstract/index.ts
@@ -1,4 +1,4 @@
 export default "";
 export {default as Portal} from "./Portal.svelte";
 export * from "./Authentication";
-export {default as Chatwoot} from "./Chatwoot.svelte";
+export * from "./Chatwoot";
