+++ clients/web/src/lib/utils/types/index.ts
@@ -1,3 +1,2 @@
 export * from "./SessionUser";
 export * from "./Config";
-export * from "./Chatwoot";
